sample_solutions/HybridSearch/api/gateway/services/auth.py
"""
Authentication Service
Handles JWT token verification from Keycloak
"""

import logging
from typing import Optional, Dict, Any
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import jwt, JWTError
import httpx
from config import settings

logger = logging.getLogger(__name__)

# OAuth2 scheme for token extraction
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token", auto_error=False)

class AuthService:
    """
    Enterprise authentication service using Keycloak JWTs.
    
    Handles validation of JWT tokens issued by Keycloak, including
    signature verification (in production) and public key caching.
    """
    
    def __init__(self):
        self.enabled = settings.deployment_phase == "production"
        self.base_url = settings.base_url
        self.realm = settings.keycloak_realm
        
        # In a real production environment, you should fetch the public key from Keycloak
        # http://{keycloak}/realms/{realm}/protocol/openid-connect/certs
        self.jwks_url = f"{self.base_url}/certs" if self.base_url else None
        self._cached_keys = None

    async def _get_public_keys(self) -> Dict[str, Any]:
        """
        Fetch and cache public keys from Keycloak.
        
        Returns:
            Dict[str, Any]: The JSON Web Key Set (JWKS) from Keycloak.
        """
        if self._cached_keys:
            return self._cached_keys
            
        if not self.jwks_url:
            return {}
            
        try:
            async with httpx.AsyncClient(verify=settings.verify_ssl) as client:
                response = await client.get(self.jwks_url)
                response.raise_for_status()
                self._cached_keys = response.json()
                return self._cached_keys
        except Exception as e:
            logger.error(f"Failed to fetch public keys from Keycloak: {e}")
            return {}

    async def verify_token(self, token: str) -> Optional[Dict[str, Any]]:
        """
        Verify the JWT token.
        
        Args:
            token (str): The raw JWT token string.
            
        Returns:
            Optional[Dict[str, Any]]: Decoded token claims if valid.
            
        Raises:
            HTTPException: If token is missing, invalid, or expired.
        """
        # BYPASS AUTH FOR TESTING
        return {"sub": "anonymous-user", "preferred_username": "dev-user", "roles": ["admin"]}
        
        if not self.enabled:
            return {"sub": "anonymous", "preferred_username": "dev-user"}
            
        if not token:
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Authentication token missing",
                headers={"WWW-Authenticate": "Bearer"},
            )
            
        try:
            # In enterprise mode, verify against provided Keycloak
            # Note: For simplicity, we are doing a loose check here.
            # In a real-world scenario, you MUST verify signatures and audience.
            
            # 1. Decode without verification to get headers/claims
            unverified_claims = jwt.get_unverified_claims(token)
            
            # 2. Add verification logic here (signatures, etc.)
            # For this PoC, we rely on the Enterprise Gateway having validated the token already
            # or we do a basic verification.
            
            return unverified_claims
            
        except JWTError as e:
            logger.error(f"JWT verification failed: {e}")
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Invalid or expired token",
                headers={"WWW-Authenticate": "Bearer"},
            )

# Dependency for FastAPI routes
async def get_current_user(token: str = Depends(oauth2_scheme)) -> Dict[str, Any]:
    """
    FastAPI dependency to get and verify the current user.
    
    Args:
        token (str): Bearer token from the Authorization header.
        
    Returns:
        Dict[str, Any]: The authenticated user's claims.
    """
    auth_service = AuthService()
    return await auth_service.verify_token(token)
