sample_solutions/HybridSearch/api/gateway/requirements.txt
# Gateway Service Requirements
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.1
aiohttp>=3.11.7

# Logging and Monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Retry Logic
tenacity==8.2.3

# Rate Limiting
slowapi==0.1.9

# Environment
python-dotenv==1.0.0

# Utilities
# Security
python-jose[cryptography]>=3.4.0
python-multipart>=0.0.22

